# Use ordered index to improve PG performances (ref #138)