ref #138

This changeset boosts insertion of records (x100 faster):

*before ~10ms per insert*

    testdb=# INSERT INTO records(user_id, resource_name, data) SELECT * FROM generate_records(10000, 1, 10);
    INSERT 0 100000
    Time: 1102111,033 ms

*after ~0.1 ms per insert*

    testdb=# INSERT INTO records(user_id, resource_name, data) SELECT * FROM generate_records(100000, 1, 10);
    INSERT 0 1000000
    Time: 109397,719 ms

@Natim @ametaireau could you give a shot locally and see if you notice the same (amazing) improvements ?

Original notes here : https://etherpad.mozilla.org/reading-list-perf-postgres