@leplatrem 2015-04-09
Quick help to review and verify locally : 

Obtain ``sample_data.sql`` from PR #191 : 

https://raw.githubusercontent.com/mozilla-services/cliquet/load_data_script/cliquet/scripts/sample_data.sql

```
postgres=# CREATE DATABASE testdb;
postgres=# \c testdb;
testdb=# \i cliquet/storage/postgresql/schema.sql
testdb=# \i sample_data.sql
testdb=# INSERT INTO records(id, user_id, resource_name, data) SELECT uuid_generate_v4(), * FROM generate_records(100000, 1, 10);


```