@Natim 2015-04-10
masterdb: Time: 809670,014 ms
branchdb: Time: 7083,479 ms

Whaou, it is incredible.